Vault MFA generator/requirements.txt
boto3==1.28.75
botocore==1.31.75
certifi==2023.7.22
charset-normalizer==3.3.0
hvac==1.2.1
idna==3.4
jmespath==1.0.1
libcst==1.1.0
loguru==0.7.2
MonkeyType==23.3.0
mypy-extensions==1.0.0
pyhcl==0.4.5
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.0.7
